Refactor: Extract join search strategies into separate modules
Split the monolithic search.rs (909 lines) into focused modules:
- search/mod.rs: Public API and core types
- search/cost.rs: Cost estimation logic
- search/dfs.rs: Sequential DFS strategy
- search/bfs.rs: Parallel BFS strategy
- search/greedy.rs: Greedy heuristic

Benefits:
- Clear separation of concerns by search strategy
- Each strategy independently testable
- Easier to add new search strategies
- Improved code organization

All tests pass, no performance regression.

Closes #2186
